+++ reikna/__init__.py
@@ -1,1 +1,2 @@
+from reikna.cluda.array_helpers import concatenate