+++ reikna/cluda/array_helpers.py
@@ -64,3 +64,60 @@ def get_method(array):
         setitem_computation, Type.from_value(temp), Type.from_value(array))
     comp(temp, array)
     return temp.get()
+
+
+def is_shape_compatible(template_shape, shape, axis):
+    for i in range(len(template_shape)):
+        if i != axis and shape[i] != template_shape[i]:
+            return False
+    return True
+
+
+def concatenate(arrays, axis=0, out=None):
+    """
+    Concatenate an iterable of arrays along ``axis`` and write them to ``out``
+    (allocating it if it is set to ``None``).
+
+    Works analogously to ``numpy.concatenate()`` (except ``axis=None`` is not supported).
+    """
+
+    # TODO: support axis=None.
+    # Requires Array.ravel() returnign an Array instead of CUDA/PyOpenCL array.
+
+    if len(arrays) == 0:
+        raise ValueError("Need at least one array to concatenate")
+    if any(array.dtype != arrays[0].dtype for array in arrays[1:]):
+        raise ValueError("Data types of all arrays must be the same")
+
+    dtype = arrays[0].dtype
+    thread = arrays[0].thread
+
+    template_shape = arrays[0].shape
+    axis = axis % len(template_shape)
+    for array in arrays[1:]:
+        if not is_shape_compatible(template_shape, array.shape, axis):
+            raise ValueError(
+                "Shapes are not compatible: " + str(template_shape) + " and " + str(shape))
+
+    out_shape = list(template_shape)
+    out_shape[axis] = sum(array.shape[axis] for array in arrays)
+    out_shape = tuple(out_shape)
+
+    if out is None:
+        out = thread.array(out_shape, dtype)
+    else:
+        if out.shape != out_shape:
+            raise ValueError(
+                "Incorrect output shape: expected " + str(out_shape) + ", got " + str(out.shape))
+        if out.dtype != dtype:
+            raise ValueError(
+                "Incorrect output dtype: expected " + str(dtype) + ", got " + str(out.dtype))
+
+    offset = 0
+    slices = [slice(None) for i in range(len(out_shape))]
+    for array in arrays:
+        slices[axis] = slice(offset, offset + array.shape[axis])
+        out[tuple(slices)] = array
+        offset += array.shape[axis]
+
+    return out