+++ doc/source/history.rst
@@ -22,6 +22,8 @@ Release history
 
 * ADDED: support for ``get()`` method for non-contiguous arrays.
 
+* ADDED: :py:func:`~reikna.concatenate` for concatenating GPU arrays.
+
 * FIXED: updated the CUDA backend for the change ``async`` -> ``async_`` in the new versions of PyCUDA. Bumped PyOpenCL and PyCUDA versions to 2018.1.1.
 
 * FIXED: an error in the conversion of `numpy.int64` to ctype for Windows.